chore(walarchive): make fadvise platform-specific
Resolve go vet errors by splitting fadvise implementation into
platform-specific files using build tags for Linux and other platforms.
