+++ pkg/walarchive/fadvise_linux_test.go
@@ -1,3 +1,5 @@
+//go:build linux
+
 /*
 Copyright The CloudNativePG Contributors
